requirements.txt
certifi==2020.6.20
chardet==3.0.4
dateparser==0.7.6
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
nested-lookup==0.2.21
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-cache==0.5.2
six==1.15.0
text-unidecode==1.3
tzlocal==2.1
urllib3==1.25.10
xmltodict==0.12.0
